git-svn-id: http://posttool.com/svn/posttool/PSWebV1.0@1933 241bf0be-db3f-0410-a781-e906b5fc445c

+++ src/com/pagesociety/web/module/ecommerce/gateway/BillingGatewayException.java
@@ -3,13 +3,13 @@ package com.pagesociety.web.module.ecommerce.gateway;
 @SuppressWarnings("serial")
 public class BillingGatewayException extends Exception 
 {
-	private float amount;
+	private double amount;
 	public BillingGatewayException(String msg)
 	{
 		super(msg);
 	}
 	
-	public BillingGatewayException(String msg,float amount)
+	public BillingGatewayException(String msg,double amount)
 	{
 		super(msg);
 		this.amount = amount;
@@ -20,7 +20,7 @@ public class BillingGatewayException extends Exception
 		super(msg, e);
 	}
 	
-	public float getAmount()
+	public double getAmount()
 	{
 		return amount;
 	}
